Update CodingReview Chapter3
