+++ Select2/src/main/java/chapter/three/Exercise1.java
@@ -1,4 +1,6 @@
-package chapter.three.exercise1;
+package chapter.three;
+
+import java.util.Random;
 
 public class Exercise1 {
     public static void main(String[] args) {
@@ -23,5 +25,31 @@ public class Exercise1 {
 
     private static boolean isSnakeEyes(Dice firstDice, Dice secondDice) {
         return firstDice.roll() == 1 && secondDice.roll() == 1;
+    }
+}
+class Dice {
+    private final int maxValue;
+
+    public Dice(int maxValue) {
+        this.maxValue = maxValue;
+        classInvariant();
+    }
+
+    private void classInvariant() {
+        if (this.maxValue < 1) {
+            throw new IllegalArgumentException("양수가 아닙니다.");
+        }
+    }
+
+    public int roll() {
+        Random random = new Random();
+        return random.nextInt(this.maxValue) + 1;
+    }
+
+    @Override
+    public String toString() {
+        return "Dice{" +
+                "maxValue=" + maxValue +
+                '}';
     }
 }
