--- Select2/src/main/java/chapter/three/exercise1/Dice.java
@@ -1,30 +1,0 @@
-package chapter.three.exercise1;
-
-import java.util.Random;
-
-public class Dice {
-    private final int maxValue;
-
-    public Dice(int maxValue) {
-        this.maxValue = maxValue;
-        classInvariant();
-    }
-
-    private void classInvariant() {
-        if (this.maxValue < 1) {
-            throw new IllegalArgumentException("양수가 아닙니다.");
-        }
-    }
-
-    public int roll() {
-        Random random = new Random();
-        return random.nextInt(this.maxValue) + 1;
-    }
-
-    @Override
-    public String toString() {
-        return "Dice{" +
-                "maxValue=" + maxValue +
-                '}';
-    }
-}
