+++ Select2/src/main/resources/sales.dat
@@ -1,0 +1,3 @@
+San Francisco: 19887.32
+Chicago:  no report received
+New York: 298734.12
